% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commit_sql.R
\name{commit_sql}
\alias{commit_sql}
\title{Commit sql from file or character variable}
\usage{
commit_sql(conn, code, is_file = TRUE)
}
\arguments{
\item{conn}{DBIConnection to sql server}

\item{code}{Character to path to sql code}

\item{is_file}{Logical indicates the code parameter refers to file name}
}
\value{
List of data frame from the last select statement, empty if no select statment
}
\description{
Commit sql from file or character variable
}
\examples{
\dontrun{
commit_sql(conn, "/home/user/bearhunt.sql")
commit_sql(conn, "select * from session.ca", is_file = F)
}
}
