% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commit_sql.R
\name{commit_sql}
\alias{commit_sql}
\title{Commit sql from file or character variable}
\usage{
commit_sql(conn, code, is_file = TRUE)
}
\arguments{
\item{conn}{DBIConnection to sql server}

\item{code}{Character of code or path to sql code}

\item{is_file}{Logical variable indicates the code parameter refers to file name, defaults to TRUE}
}
\value{
List of data frames from all select statement, empty if no select statment
}
\description{
Commit sql from file or character variable
}
\examples{
\dontrun{
jcc <- RJDBC::JDBC("com.ibm.db2.jcc.DB2Driver", "C:/db2jcc4.jar")
conn <- RJDBC::dbConnect(jcc, "jdbc:db2://hostname:12345/dbname", "user", "pwd")
commit_sql(conn, "/home/user/bearhunt.sql")
commit_sql(conn, "select * from session.ca", is_file = F)
}
}
\seealso{
\url{https://github.com/dajuntian/sqlTXT}
}
